Move User role handlers to basic permmisons
Handlers using basic permissions objects need to be inside basic
permissions module, so we won't need imports inside functions.
